tests/real-agent/requirements.txt
# Requirements for real GitHub agent with OpenAI and TameSDK
openai>=1.40.0
mcp>=1.0.0
httpx>=0.25.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0